VM: Fix problem with trivial ranges (patch #10362).
* libinterp/octave-value/ov-range.cc, libinterp/octave-value/ov-range.h
(ov_range<double>::could_be_trivial_range): Check that double precision values
can be represented as `int` data type.

* test/compile/bytecode.tst, test/compile/bytecode_for.m: Add tests with more
for-loop expressions.

+++ libinterp/octave-value/ov-range.h
@@ -56,7 +56,7 @@ class octave_trivial_range : public octave_base_value
 {
 public:
 
-  octave_trivial_range (octave_idx_type numel, double base, double incr)
+  octave_trivial_range (octave_idx_type numel, int base, int incr)
       : m_numel (numel), m_base (base), m_increment(incr) { }
 
   octave_trivial_range () {};
